+++ examples/theme.rs
@@ -1,0 +1,40 @@
+#![allow(clippy::single_match)]
+
+use simple_logger::SimpleLogger;
+use winit::{
+    event::{Event, WindowEvent},
+    event_loop::{ControlFlow, EventLoop},
+    window::{Theme, WindowBuilder},
+};
+
+fn main() {
+    SimpleLogger::new().init().unwrap();
+    let event_loop = EventLoop::new();
+
+    let window = WindowBuilder::new()
+        .with_title("A fantastic window!")
+        .with_theme(Some(Theme::Dark))
+        .build(&event_loop)
+        .unwrap();
+
+    println!("Initial theme: {:?}", window.theme());
+
+    event_loop.run(move |event, _, control_flow| {
+        *control_flow = ControlFlow::Wait;
+
+        match event {
+            Event::WindowEvent {
+                event: WindowEvent::CloseRequested,
+                ..
+            } => *control_flow = ControlFlow::Exit,
+            Event::WindowEvent {
+                event: WindowEvent::ThemeChanged(theme),
+                window_id,
+                ..
+            } if window_id == window.id() => {
+                println!("Theme is changed: {:?}", theme)
+            }
+            _ => (),
+        }
+    });
+}